Module hyperplonk::prelude

Re-exports

pub use crate::HyperPlonkSNARK;

Structs

CustomizedGates

Customized gate is a list of tuples of (coefficient, selector_index, wire_indices)

MockCircuit

SelectorColumn

A column of selectors of length #constraints

WitnessColumn

A column of witnesses of length #constraints

Enums

HyperPlonkErrors

A enum specifying the possible failure modes of hyperplonk.