hyperplonk
Trait HyperPlonkSNARK

pub trait HyperPlonkSNARK<E, PCS>: PermutationCheck<E, PCS>
where
    E: Pairing,
    PCS: PolynomialCommitmentScheme<E>,
{
    type Index;
    type ProvingKey;
    type VerifyingKey;
    type Proof;

    // Required methods
    fn preprocess(
        index: &Self::Index,
        pcs_srs: &PCS::SRS,
    ) -> Result<(Self::ProvingKey, Self::VerifyingKey), HyperPlonkErrors>;
    fn prove(
        pk: &Self::ProvingKey,
        pub_input: &[E::ScalarField],
        witnesses: &[WitnessColumn<E::ScalarField>],
    ) -> Result<Self::Proof, HyperPlonkErrors>;
    fn verify(
        vk: &Self::VerifyingKey,
        pub_input: &[E::ScalarField],
        proof: &Self::Proof,
    ) -> Result<bool, HyperPlonkErrors>;
}

A trait for HyperPlonk SNARKs. A HyperPlonk is derived from ZeroChecks and PermutationChecks.

Required Associated Types

type Index

type ProvingKey

type VerifyingKey

type Proof

Required Methods

fn preprocess( index: &Self::Index, pcs_srs: &PCS::SRS, ) -> Result<(Self::ProvingKey, Self::VerifyingKey), HyperPlonkErrors>

Generate the preprocessed polynomials output by the indexer.

Inputs:

index: HyperPlonk index

pcs_srs: Polynomial commitment structured reference string

Outputs:

The HyperPlonk proving key, which includes the preprocessed polynomials.

The HyperPlonk verifying key, which includes the preprocessed polynomial commitments

fn prove( pk: &Self::ProvingKey, pub_input: &[E::ScalarField], witnesses: &[WitnessColumn<E::ScalarField>], ) -> Result<Self::Proof, HyperPlonkErrors>

Generate HyperPlonk SNARK proof.

Inputs:

pk: circuit proving key

pub_input: online public input

witness: witness assignment

Outputs:

The HyperPlonk SNARK proof.

fn verify( vk: &Self::VerifyingKey, pub_input: &[E::ScalarField], proof: &Self::Proof, ) -> Result<bool, HyperPlonkErrors>

Verify the HyperPlonk proof.

Inputs:

vk: verifying key

pub_input: online public input

proof: HyperPlonk SNARK proof challenges

Outputs:

Return a boolean on whether the verification is successful

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.