Trait hyperplonk::HyperPlonkSNARK

pub trait HyperPlonkSNARK<E, PCS>: SumCheck<E::Fr> + ZeroCheck<E::Fr> + PermutationCheck<E::Fr> where
    E: PairingEngine,
    PCS: PolynomialCommitmentScheme<E>, {
    type Parameters;
    type ProvingKey;
    type VerifyingKey;
    type Proof;
    type SubClaim;

    fn preprocess(
        params: &Self::Parameters, 
        pcs_srs: &PCS::SRS, 
        permutation: &[E::Fr], 
        selectors: &[SelectorColumn<E::Fr>]
    ) -> Result<(Self::ProvingKey, Self::VerifyingKey), HyperPlonkErrors>;
    fn prove(
        pk: &Self::ProvingKey, 
        pub_input: &[E::Fr], 
        witnesses: &[WitnessColumn<E::Fr>], 
        transcript: &mut Self::Transcript
    ) -> Result<Self::Proof, HyperPlonkErrors>;
    fn verify(
        params: &Self::VerifyingKey, 
        pub_input: &[E::Fr], 
        proof: &Self::Proof, 
        transcript: &mut Self::Transcript
    ) -> Result<Self::SubClaim, HyperPlonkErrors>;
}

A trait for HyperPlonk Poly-IOPs. A HyperPlonk is derived from SumChecks, ZeroChecks and PermutationChecks.

Required Associated Types

type Parameters

type ProvingKey

type VerifyingKey

type Proof

type SubClaim

Required Methods

fn preprocess(
params: &Self::Parameters,
pcs_srs: &PCS::SRS,
permutation: &[E::Fr],
selectors: &[SelectorColumn<E::Fr>]
) -> Result<(Self::ProvingKey, Self::VerifyingKey), HyperPlonkErrors>

Generate the preprocessed polynomials output by the indexer.

Inputs:

params: HyperPlonk instance parameters

permutation: the permutation for the copy constraints

selectors: the list of selector vectors for custom gates Outputs:

The HyperPlonk proving key, which includes the preprocessed polynomials.

fn prove(
pk: &Self::ProvingKey,
pub_input: &[E::Fr],
witnesses: &[WitnessColumn<E::Fr>],
transcript: &mut Self::Transcript
) -> Result<Self::Proof, HyperPlonkErrors>

Generate HyperPlonk SNARK proof.

Inputs:

pk: circuit proving key

pub_input: online public input

witness: witness assignment

transcript: the transcript used for generating pseudorandom challenges Outputs:

The HyperPlonk SNARK proof.

fn verify(
params: &Self::VerifyingKey,
pub_input: &[E::Fr],
proof: &Self::Proof,
transcript: &mut Self::Transcript
) -> Result<Self::SubClaim, HyperPlonkErrors>

Verify the HyperPlonk proof and generate the evaluation subclaims to be checked later by the SNARK verifier.

Inputs:

params: instance parameter

pub_input: online public input

proof: HyperPlonk SNARK proof

transcript: the transcript used for generating pseudorandom challenges Outputs:

Return error if the verification fails, otherwise return the evaluation subclaim